Move set and get to utilities
